delete createLogger
